+++ alembic.ini
@@ -13,7 +13,7 @@ script_location = %(here)s/src/models/migrations
 # file_template = %%(year)d_%%(month).2d_%%(day).2d_%%(hour).2d%%(minute).2d-%%(rev)s_%%(slug)s
 
 # sys.path path, will be prepended to sys.path if present.
-# defaults to the current working directory.  for multiple paths, the path separator
+# defaults to the current working directory. for multiple paths, the path separator
 # is defined by "path_separator" below.
 prepend_sys_path = .
 
@@ -83,7 +83,6 @@ path_separator = os
 # database URL.  This is consumed by the user-maintained env.py script only.
 # other means of configuring database URLs may be customized within the env.py
 # file.
-# src/modelsディレクトリから見た相対パスでデータベースを指定
 sqlalchemy.url = sqlite:///storage/data/tasks.db
