+++ src/models/__init__.py
@@ -2,17 +2,18 @@
 # モデル層の初期化ファイル
 
 # [AI GENERATED] 新しいモデルのインポート
-from models.new_task import Task, TaskCreate, TaskRead, TaskStatus, TaskUpdate
 from models.project import Project, ProjectCreate, ProjectRead, ProjectStatus, ProjectUpdate
 from models.tag import Tag, TagCreate, TagRead, TagUpdate
-from models.task import OldTask, OldTaskCreate, OldTaskRead, OldTaskUpdate
+from models.task import (
+    Task,
+    TaskCreate,
+    TaskRead,
+    TaskStatus,
+    TaskUpdate,
+)
 from models.task_tag import TaskTag, TaskTagCreate, TaskTagRead
 
 __all__ = [
-    "OldTask",
-    "OldTaskCreate",
-    "OldTaskRead",
-    "OldTaskUpdate",
     "Project",
     "ProjectCreate",
     "ProjectRead",
